pyproject.toml
[project]
name = "betting-recommender"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "catboost>=1.2.8",
    "feast>=0.49.0",
    "joblib>=1.5.1",
    "mlflow>=2.22.0",
    "mysql-connector-python>=9.3.0",
    "pymysql>=1.1.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "sqlalchemy>=2.0.41",
]
